RESET LockStepper code and tests.  look to Vtrace PR#406 for these changes
